% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_gkrls_me}
\alias{cpp_gkrls_me}
\title{Internal C++ function to calculate marginal effects}
\usage{
cpp_gkrls_me(
  std_X_train,
  std_X_test,
  bandwidth,
  family,
  mahal,
  sd_y,
  offset,
  any_Z,
  tZ,
  tS,
  fe_mean,
  re_mean,
  SIZE_PARAMETER,
  vcov_ridge,
  FE_matrix_test,
  W_Matrix,
  WX_test,
  WX_train,
  raw_X_test,
  std_mean,
  std_whiten,
  type_mfx,
  fd_matrix,
  std_fd_matrix
)
}
\description{
Internal C++ function to calculate marginal effects
}
