% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_sketched_kernel}
\alias{create_sketched_kernel}
\title{Create the sketched kernel}
\usage{
create_sketched_kernel(X_test, X_train, S, bandwidth, raw = FALSE)
}
\arguments{
\item{X_test}{Test data}

\item{X_train}{Train data}

\item{S}{Sketch matrix}

\item{bandwidth}{Kernel bandwidth}
}
\description{
Create the sketched kernel
}
\keyword{internal}
