% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_function.R
\name{gKRLS_gam_control}
\alias{gKRLS_gam_control}
\title{Control options for Kernel-Gam}
\usage{
gKRLS_gam_control(
  truncate.eigen.tol = sqrt(.Machine$double.eps),
  demean_kernel = FALSE,
  sketch_method = "nystrom",
  no.rescale = FALSE,
  standardize = "scaled",
  sketch_size = function(N) {     ceiling(N^(1/3)) * 5 },
  remove_instability = TRUE
)
}
\description{
Control options for Kernel-Gam
}
