% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_mgcv_functions.R
\name{internal_functions}
\alias{internal_functions}
\alias{internal_ocat_jacobian}
\alias{internal_ziP_jacobian}
\alias{internal_multinom_jacobian}
\title{Internal Functions for Predict}
\usage{
internal_ocat_jacobian(
  family,
  se = FALSE,
  eta = NULL,
  y = NULL,
  X = NULL,
  beta = NULL,
  off = NULL,
  Vb = NULL
)

internal_ziP_jacobian(
  family,
  se = FALSE,
  eta = NULL,
  y = NULL,
  X = NULL,
  beta = NULL,
  off = NULL,
  Vb = NULL
)

internal_multinom_jacobian(
  family,
  se = FALSE,
  eta = NULL,
  y = NULL,
  X = NULL,
  beta = NULL,
  off = NULL,
  Vb = NULL
)
}
\arguments{
\item{family}{Name of family}

\item{se}{Calculate standard errors}

\item{eta}{Not used internally for \code{gKRLS}}

\item{y}{Not used internally for \code{gKRLS}}

\item{X}{Design matrix}

\item{beta}{Coefficients}

\item{off}{Not used internally for \code{gKRLS}}

\item{Vb}{Variance matrix}
}
\description{
These functions are taken from "mgcv" with slight modifications, mostly to
export the derivative of the predicted value with respect to the linear
predictor so they can be used to calculate the jacobian for the delta method.
Please see \code{mgcv} for more details on the argument.
}
\keyword{internal}
