% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfx_function.R
\name{legacy_marginal_effect}
\alias{legacy_marginal_effect}
\title{Marginal Effects by Derivative}
\usage{
legacy_marginal_effect(object, newdata, keep = NULL, method = "cpp")
}
\description{
Find the (average) marginal effect and standard error by taking the
analytical derivative of the conditional expectation function. This method is
designed to provide comparability with `KRLS` but is limited in the scenarios
where it can be applied. Specifically, it can only compute effects with a
single kernel and where variables are not both in the kernel and outside the
kernel.
}
\details{
Because of these restrictions, users should rely on `calculate_effects` and
this function may be removed in future updates. If a derivative is really
desired, this can be approximated as described in the `calculate_effects`
documentation.
}
