% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effect.R
\name{marginal_effect_base}
\alias{marginal_effect_base}
\title{Estimate marginal effects}
\usage{
marginal_effect_base(
  object,
  newdata,
  newkernel_X = NULL,
  keep = NULL,
  method = "cpp"
)
}
\description{
Estimate marginal effects
}
