% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minibatch_kmeans_cpp}
\alias{minibatch_kmeans_cpp}
\title{Mini-batch k-means clustering for Nystrom landmark selection}
\usage{
minibatch_kmeans_cpp(X, k, batch_size, max_iter, tol)
}
\arguments{
\item{X}{Data matrix (N x P)}

\item{k}{Number of clusters (landmark points)}

\item{batch_size}{Size of each mini-batch}

\item{max_iter}{Maximum number of iterations}

\item{tol}{Convergence tolerance on centroid shift}
}
\value{
Matrix of k centroids (k x P)
}
\description{
Uses k-means++ initialization and vectorized batch distance computation
via Eigen for fast landmark point selection.
}
\keyword{internal}
