% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_function.R
\name{predict.gKRLS}
\alias{predict.gKRLS}
\title{Prediction after gKRLS
Pass "newdata"}
\usage{
\method{predict}{gKRLS}(object, newdata, type = c("link", "response"), return_full = FALSE)
}
\description{
Prediction after gKRLS
Pass "newdata"
}
