% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_function.R
\name{Predict.matrix.gKRLS.smooth}
\alias{Predict.matrix.gKRLS.smooth}
\title{Predict Methods for gKRLS smooth}
\usage{
\method{Predict.matrix}{gKRLS.smooth}(object, data)
}
\arguments{
\item{object}{a smooth object; see documentation for other methods in
\code{mgcv}.}

\item{data}{a data.frame; see documentation for other methods in
\code{mgcv}.}
}
\description{
Predict Methods for gKRLS smooth
}
\keyword{internal}
