% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_function.R
\name{smooth.construct.gKRLS.smooth.spec}
\alias{smooth.construct.gKRLS.smooth.spec}
\title{Constructor for gKRLS smooth}
\usage{
\method{smooth.construct}{gKRLS.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth object; see documentation for other methods in
\code{mgcv}.}

\item{data}{a data.frame; see documentation for other methods in
\code{mgcv}.}

\item{knots}{not used}
}
\description{
Constructor for gKRLS smooth
}
\keyword{See}
\keyword{\link{gKRLS}}
\keyword{details.}
\keyword{for}
\keyword{internal}
