% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_function.R
\encoding{UTF-8}
\name{calculate_effects}
\alias{calculate_effects}
\alias{calculate_interactions}
\alias{get_individual_effects}
\alias{print.gKRLS_mfx}
\alias{summary.gKRLS_mfx}
\title{Marginal Effects}
\usage{
calculate_effects(
  model,
  data = NULL,
  variables = NULL,
  vcov = NULL,
  raw = FALSE,
  individual = FALSE,
  conditional = NULL,
  epsilon = 1e-07,
  verbose = FALSE,
  continuous_type = c("IQR", "minmax", "derivative", "onesd", "predict",
    "second_derivative")
)

calculate_interactions(
  model,
  variables,
  QOI = c("AMIE", "ACE", "AME", "AIE"),
  ...
)

get_individual_effects(x)

\method{print}{gKRLS_mfx}(x, ...)

\method{summary}{gKRLS_mfx}(object, ...)
}
\arguments{
\item{model}{A model estimated from \code{mgcv}.}

\item{data}{A new data frame that used to calculate the marginal effect, or
set to \code{NULL}, which the data used to estimate the model will be used.
The default is \code{NULL}.}

\item{variables}{A character vector that specifies the variables for which to
calculate effects. The default, \code{NULL}, calculates marginal effects
for all variables.}

\item{vcov}{A matrix that specifies the covariance matrix on the parameters.
The default, \code{NULL}, uses the standard covariance matrix from
\code{mgcv}. This can be used to specify clustered or robust matrices
using, e.g., the \code{sandwich} package.}

\item{raw}{Return the raw values used to calculate the difference in addition
to the estimated effect. Default is \code{FALSE}. If \code{TRUE}, an
additional column \code{...id} is present in the estimated effects that
reports whether the the row corresponds to the effect (\code{effect}), the
first value (\code{raw_0}) or the second value (\code{raw_1}) where
\code{effect=raw_1 - raw_0}. 

For \code{"derivative"}, this must be scaled by the step size. For
\code{"second_derivative"}, \code{effect=raw_2 - 2 * raw_1 + raw_0}, scaled
by the step size; see the discussion for \code{epsilon}.}

\item{individual}{A value of \code{TRUE} calculates individual effects (i.e.
an effect for each observation in the provided data). The default is
\code{FALSE}.}

\item{conditional}{This is an analogue of Stata's \code{at()} option and the
\code{at} argument in the \code{margins}  package. For a marginal effect on
some variable \code{"a"}, this can specify the values for other covariates,
e.g. \code{"b"}, to be held at. Examples are provided below. This should be
either \code{NULL} (default) or a data.frame where each row is a
combination of covariate values to be held fixed. Marginal effects are
calculated separately for each row of \code{conditional}.}

\item{epsilon}{A numerical value to define the step size when calculating
  numerical derivatives.  For \code{"derivative"}, the step size for the
  approximation is  \eqn{h = \epsilon \cdot \mathrm{max}(1,
  \mathrm{max}(|x|))}{h = \epsilon * max(1, max(|x|))}, i.e. \eqn{f'(x)
  \approx \frac{f(x+h) - f(x-h)}{2h}}{f'(x) ≈ [f(x+h)-f(x-h)]/(2h)}. Please
  see Leeper (2016) for more details.

  For \code{"second_derivative"}, the step size is \eqn{h = [\epsilon \cdot
  \mathrm{max}(1, \mathrm{max}(|x|))]^{0.5}}{h=[\epsilon * max(1, max(|x|))]^{0.5}}, i.e.
  \eqn{f''(x) \approx \frac{f(x+h) - 2 f(x) + f(x-h)}{h^2}}{f''(x) ≈ [f(x+h)
  - 2 f(x) + f(x-h)]/h^2}}

\item{verbose}{A logical value indicates whether to report progress when
calculating the marginal effects.}

\item{continuous_type}{A character string indicating the type of marginal
effects to estimate when the variable is continuous (i.e. not binary,
logical, factor, or character). Options are \code{"IQR"} (compares the
variable at its 25\% and 75\% percentile), \code{"minmax"} (compares the
variable at its minimum and maximum), \code{"derivative"} (numerically
approximates the derivative at each observed value),
\code{"second_derivative"} (numerically approximates the second derivative
at each observed value), \code{"onesd"} (compares one standard deviation
below and one standard deviation above). It may also accepted a \bold{named
list} where each named element corresponds to a numeric variable and has a
two-length vector as each element. The two values are then compared.

A special option (\code{"predict"}) produces predictions (e.g.,
\code{predict(model, ..., type = "response")}) at each observed value and
then averages them together. This in conjunction with \code{conditional}
provides a way of calculating, for example, predicted probability curves
using an "observed value" approach (e.g., Hanmer and Kalkan 2013). The
examples below provide an illustration.}

\item{QOI}{Quantity of interest to calculate for
\code{calculate_interactions}. Options include AME (average marginal
effect), ACE (average combination effect), AIE (average interaction effect)
and AMIE (average marginal interaction effect); see "Details" for more
information.}

\item{...}{Additional arguments (unused).}

\item{x}{Object fit with \code{calculate_effects} or \code{legacy_marginal_effect}}

\item{object}{A model estimated from \code{mgcv}.}
}
\value{
\code{calculate_effects} returns a data.frame of class
  \code{"gKRLS_mfx"} that contains a data.frame with the estimated average
  marginal effects. \code{"type"} reports the type of marginal effect
  calculated. For families with multiple predicted outcomes (e.g.,
  multinomial), the column \code{"response"} numbers the different outcomes
  in the same order as \code{predict.gam(object)} for the specified family.
  Many (but not all) extended and generalized families from \code{mgcv} are
  included.
  
  To estimated predicted values at different covariate strata (i.e., create
  an "observed value" predicted probability curve for a logistic regression),
  please use the \code{conditional} argument while setting
  \code{continuous_type = "predict"}. Please see the examples for an
  illustration.
  
  This data.frame has attributes, listed below, that are used in other
  functions. The function \code{get_individual_effects} extracts the
  individual-level effects that are estimated if \code{individual=TRUE}.
  
  \itemize{
  \item{"jacobian": } This reports the corresponding Jacobian used to
  calculate the standard error (via the delta method) for the estimate. There
  is one row for each row in "marginal_effects". This can be used to, for
  example, calculate a standard error on the difference between two estimated
  marginal effects.
  \item{"counter": } A placeholder for the number of marginal effects calculated.
  \item{"N_eff": The number of observations (in the estimation data) minus
  the effective degrees of freedom. This is used when calculating p-values as
  the degrees of freedom for the t-distribution.}
  \item{"N": The number of observations.}
  }
 
 \code{calculate_interactions} provides some simple functions for calculate
 interaction effects between variables. The default quantities it can produce
 are listed below. Egami and Imai (2019) provide a detailed exposition of
 these quantities. All marginalization is done using an "observed value"
 approach, i.e. over the estimation data or a custom dataset provided to
 \code{data}.
 \itemize{
   \item{"AME" or Average Marginal Effect: } This is the standard quantity
   reported from \code{calculate_effects}.
   \item{"ACE" or Average Combination Effect: } This is the effect of
   changing two variables simultaneously on the outcome.
   \item{"AMIE" or Average Marginal Interaction Effect: } This is ACE minus
   each corresponding AME.
   \item{"AIE" or Average Interaction Effect:} This has a "conditional
   effect" interpretation and reports the difference in average effect of one
   variable ("A") between two different levels of a second variable ("B").
 }
}
\description{
This function calculates the marginal effects after estimating a model with
\code{gam} or \code{bam}. For continuous predictors, a numerical
approximation of the partial derivative is available following Leeper (2016).
}
\examples{
set.seed(654)
n <- 50
x1 <- rnorm(n)
x2 <- rnorm(n)
x3 <- rnorm(n)
state <- sample(letters[1:5], n, replace = TRUE)
y <- 0.3 * x1 + 0.4 * x2 + 0.5 * x3 + rnorm(n)
data <- data.frame(y, x1, x2, x3, state)

# Make character variables into factors for mgcv
data$state <- factor(data$state)

# A gKRLS model
fit_gKRLS <- mgcv::gam(y ~ state + s(x1, x2, x3, bs = "gKRLS"), data = data)

# calculate marginal effect using derivative
calculate_effects(fit_gKRLS, variables = "x1", continuous_type = "derivative")

# calculate marginal effect by specifying conditional variables
calculate_effects(fit_gKRLS,
  variables = "x1",
  conditional = data.frame(x2 = c(0.6, 0.8), x3 = 0.3)
)

# calculate interaction effects between two variables
# use the default setting ("IQR") for the baseline and
# comparison categories for each variable
calculate_interactions(fit_gKRLS, 
   variables = list(c("x1", "x2")),
   QOI = c('AIE', 'AMIE')
)

# calculate marginal effect by specifying a factor conditional variable
# estimate the individual marginal effects
out <- calculate_effects(fit_gKRLS,
  variables = "x1", individual = TRUE,
  conditional = data.frame(state = c("a", "b", "c")), continuous_type = "derivative"
)

# Extract the individual marginal effects: 
# shorthand for attr(fit_main, 'individual')
get_individual_effects(out)

# calculated the average expected value across a grid of "x1" 
# using an observed value approach for the other covariates
calculate_effects(fit_gKRLS, conditional = data.frame(x1 = c(0, 0.2, 0.4, 0.6)),
  continuous_type = 'predict'
)
}
\references{
Egami, Naoki and Kosuke Imai. 2019. "Causal Interaction in Factorial
Experiments: Application to Conjoint Analysis." \emph{Journal of the American
Statistical Association}. 114(526):529-540.

Hanmer, Michael J. and Kerem Ozan Kalkan. 2013. "Behind the Curve: Clarifying
the Best Approach to Calculating Predicted Probabilities and Marginal Effects
from Limited Dependent Variable Models." \emph{American Journal of Political
Science} 57(1): 263-277.

Leeper, Thomas J. 2016. "Interpreting Regression Results using Average
Marginal Effects with R's \code{margins}." Working paper available at
\url{https://s3.us-east-2.amazonaws.com/tjl-sharing/assets/AverageMarginalEffects.pdf}.
}
