% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AME_function.R
\name{calculate_effects}
\alias{calculate_effects}
\alias{print.gKRLS_mfx}
\alias{summary.gKRLS_mfx}
\title{Marginal Effects}
\usage{
calculate_effects(
  model,
  data = NULL,
  variables = NULL,
  vcov = NULL,
  raw = FALSE,
  individual = FALSE,
  conditional = NULL,
  epsilon = 1e-07,
  verbose = FALSE,
  continuous_type = c("IQR", "minmax", "derivative", "onesd", "predict",
    "second_derivative")
)

\method{print}{gKRLS_mfx}(x, ...)

\method{summary}{gKRLS_mfx}(object, ...)
}
\arguments{
\item{model}{A model estimated from \code{mgcv}.}

\item{data}{A new data frame that used to calculate the marginal effect, or
set to \code{NULL}, which the data used to estimate the model will be used.
The default is \code{NULL}.}

\item{variables}{A character vector that specifies the variables for which to
calculate effects. The default, \code{NULL}, calculates marginal effects
for all variables.}

\item{vcov}{A matrix that specifies the covariance matrix on the parameters.
The default, \code{NULL}, uses the standard covariance matrix from
\code{mgcv}. This can be used to specify clustered or robust matrices
using, e.g., the \code{sandwich} package.}

\item{raw}{Argument used for internal functions only. Default is \code{FALSE}.}

\item{individual}{A value of \code{TRUE} calculates individual effects (i.e.
an effect for each observation in the provided data). The default is
\code{FALSE}.}

\item{conditional}{This is an analogue of Stata's \code{at()} option and the
\code{at} argument in the \code{margins}  package. For a marginal effect on
some variable \code{"a"}, this can specify the values for other covariates,
e.g. \code{"b"}, to be held at. Examples are provided below. This should be
either \code{NULL} (default) or a data.frame where one marginal effect (per
variable) is provided for each row of \code{conditional}.}

\item{epsilon}{A numerical value to define the step when calculating
numerical derivatives. See Leeper (2016) for details.}

\item{verbose}{A logical value indicates whether to report progress when
calculating the marginal effects.}

\item{continuous_type}{A character string indicating the type of marginal
effects to estimate when the variable is continuous (i.e. not binary,
logical, factor, or character). Options are \code{"IQR"} (compares the
variable at its 25\% and 75\% percentile), \code{"minmax"} (compares the
variable at its minimum and maximum), \code{"derivative"} (numerically
approximates the derivative at each observed value),
\code{"second_derivative"} (numerically approximates the second derivative
at each observed value), \code{"onesd"} (compares one standard deviation
below and one standard deviation above). It may also accepted a \bold{named
list} where each named element corresponds to a numeric variable and has a
two-length vector as each element. The two values are then compared.

A special option (\code{"predict"}) produces predictions (e.g.,
\code{predict(model, ..., type = "response")}) at each observed value and
then averages them together. This in conjunction with \code{conditional}
provides a way of calculating, for example, predicted probability curves
using an "observed value" approach (e.g., Hanmer and Kalkan 2013). The
examples below provide an illustration.}

\item{x}{Object fit with \code{calculate_effects} or \code{legacy_marginal_effect}}

\item{...}{Additional arguments (unused).}

\item{object}{A model estimated from \code{mgcv}.}
}
\value{
\code{calculate_effects} returns a list of class \code{"gKRLS_mfx"}
  that contains the following elements.
  \itemize{
  \item{"marginal_effects": } A data.frame containing the estimated marginal
  effects. \code{"type"} reports the type of marginal effect calculated. The
  estimates, standard errors, t-statistics, and p-values are also reported.
  \item{"jacobian": } This reports the corresponding Jacobian used to
  calculate the standard error (via the delta method) for the estimate. There
  is one row for each row in "marginal_effects". This can be used to, for
  example, calculate a standard error on the difference between two estimated
  marginal effects.
  \item{"counter": } A placeholder for the number of marginal effects calculated.
  \item{"N_eff": The number of observations (in the estimation data) minus
  the effective degrees of freedom. This is used when calculating p-values as
  the degrees of freedom for the t-distribution.}
  \item{"N": The number of observations.}
  }
}
\description{
This function calculates the marginal effects after estimating a model with
\code{gam} or \code{bam}. For continuous predictors, a numerical
approximation of the partial derivative is available following Leeper (2016).
}
\examples{
n <- 50
x1 <- rnorm(n)
x2 <- rnorm(n)
x3 <- rnorm(n)
state <- sample(letters[1:5], n, replace = TRUE)
y <- 0.3 * x1 + 0.4 * x2 + 0.5 * x3 + rnorm(n)
data <- data.frame(y, x1, x2, x3, state)

# Make character variables into factors for mgcv
data$state <- factor(data$state)

# A gKRLS model
gkrls_est <- mgcv::gam(y ~ state + s(x1, x2, x3, bs = "gKRLS"), data = data)

# calculate marginal effect using derivative
calculate_effects(gkrls_est, variables = "x1", continuous_type = "derivative")

# calculate marginal effect by specifying conditional variables
calculate_effects(gkrls_est,
  variables = "x1",
  conditional = data.frame(x2 = c(0.6, 0.8), x3 = 0.3)
)

# calculate marginal effect by specifying a factor conditional variable
calculate_effects(gkrls_est,
  variables = "x1",
  conditional = data.frame(state = c("a", "b", "c")), continuous_type = "derivative"
)
# calculated the average expected value across a grid of "x1" using an observed value approach
calculate_effects(gkrls_est, conditional = data.frame(x1 = c(0, 0.2, 0.4, 0.6)),
  continuous_type = 'predict'
)
}
\references{
Hanmer, Michael J. and Kerem Ozan Kalkan. 2013. "Behind the Curve: Clarifying
the Best Approach to Calculating Predicted Probabilities and Marginal Effects
from Limited Dependent Variable Models." \emph{American Journal of Political
Science} 57(1): 263-277.

Leeper, Thomas J. 2016. "Interpreting Regression Results using Average
Marginal Effects with R's \code{margins}." Working paper available at
\url{https://s3.us-east-2.amazonaws.com/tjl-sharing/assets/AverageMarginalEffects.pdf}.
}
