% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{gKRLS}
\alias{gKRLS}
\title{Fit a generalized KRLS model using Laplace approximation}
\usage{
gKRLS(
  y,
  X,
  family,
  sketch_size = ceiling(nrow(X)^(1/3)) * 5,
  sketch_method = "gaussian",
  sketch_prob = NULL,
  bandwidth = NULL,
  standardize = "Mahalanobis",
  control = list(init_var = 1, truncate_eigen = sqrt(.Machine$double.eps)),
  intercept = FALSE
)
}
\description{
Fit a generalized KRLS model using Laplace approximation
}
