% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{gKRLS}
\alias{gKRLS}
\title{gKRLS}
\usage{
gKRLS(
  truncate.eigen.tol = sqrt(.Machine$double.eps),
  demean_kernel = FALSE,
  sketch_method = "nystrom",
  no.rescale = FALSE,
  standardize = "scaled",
  sketch_size = function(N) {     ceiling(N^(1/3)) * 5 },
  remove_instability = TRUE
)
}
\arguments{
\item{sketch_method}{String vector that specifies which kernel sketch method
should be used. Options include ``gaussian'': gaussian kernel
approximation, ``nystrom'': , bernoulli, and no sketch. The default is
gaussian.}

\item{standardize}{A string vector that specifies which standardization
method should be used. Must be one of scaled, Mahalanobis, or none, which
menas no standardization. The default is ``Mahalanobis''.}

\item{sketch_size}{Specify the dimension of the randomized sketches of the
kernel matrix. The default size is ``ceiling(nrow(X)^(1/3)) * 5,'' where X
is matrix of the independent variable.}

\item{data}{data.frame containing the outcome and any fixed effects}

\item{kernel_X}{matrix containing the variables used to build the kernel}

\item{X}{A matrix of numeric observations of independent variables. Factors
and missing values not allowed. Also, no intercept is required.}

\item{family}{Provide a family as in the Generalized Linear Models (GLM). For
example: gaussian or poisson.}

\item{bandwidth}{A numerical number indicating the bandwidth for kernel. The
default is  ``2*ncol(X).''}

\item{control}{A list containing control parameters. The truncate_eigen determines how
much eigenvalues should be truncated. If truncate_eigen set to 1e-6, this
means we only keep eigenvalue greater or equal to 1e-6. The default is
sqrt(.Machine$double.eps), where .Machine$double.eps is the smallest
positive floating-point number x such that 1 + x != 1.}
}
\description{
Fit a generalized KRLS model using mgcv
}
\details{
The function `gKRLS` should be given as a control argument to any relevant
function for fitting a generalized additive model from `mgcv` (e.g. `bam`,
`gam`, `gamm4`). Its arguments are described below with a simple example.

A kernel can be specified by `s(..., bs = "gKRLS")`.

NOTE: Variables must be separated by commas inside of the `s(..)`.
}
