% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{gKRLS}
\alias{gKRLS}
\title{Fit a generalized KRLS model using Laplace approximation}
\usage{
gKRLS(
  y,
  formula,
  data,
  kernel_X,
  family,
  sketch_size = ceiling(nrow(kernel_X)^(1/3)) * 5,
  sketch_method = "gaussian",
  sketch_prob = NULL,
  bandwidth = NULL,
  standardize = "scaled",
  control = list(init_var = 1, intercept = TRUE, truncate_eigen =
    sqrt(.Machine$double.eps))
)
}
\description{
Fit a generalized KRLS model using Laplace approximation
}
