% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_function.R
\name{SL.mgcv}
\alias{SL.mgcv}
\alias{predict.SL.mgcv}
\alias{add_bam_to_mlr3}
\title{Machine Learning with gKRLS}
\usage{
SL.mgcv(Y, X, newX, formula, family, id, obsWeights, bam = FALSE, ...)

\method{predict}{SL.mgcv}(object, newdata, allow_missing_levels = TRUE, ...)

add_bam_to_mlr3()
}
\description{
This provides a number of functions to integrate machine learning with gKRLS
(and more generally `mgcv`). Integrations into `SuperLearner` and `DoubleML`
(via `mlr3`) are provided below.
}
\details{
SuperLearner integration is provided by `SL.mgcv` and the corresponding
predict method. `mgcv::bam` can be enabled by using `bam = TRUE`. Note that a
formula must be explicitly provided as a *character*, e.g. " ~ X1 + X2".

`DoubleML` integration is provided in two ways. First, one could load
`mlr3extralearners` to access `regr.gam` and `classif.gam`. Second, this
package provides `mgcv::bam` integration directly with a slight adaption of
the `mlr3extralearner` implementation. These can be either manually added to
the list of `mlr3` learners by calling `add_bam_to_mlr3()` or direct usage.
Examples are provided below. For `classif.bam` and `regr.bam`, the formula
argument is mandatory.
}
